fixing minor bug

+++ src/domain/answers/answers_dto.go
@@ -3,10 +3,10 @@ package answers
 import "github.com/dzikrisyafi/kursusvirtual_utils-go/rest_errors"
 
 type Answer struct {
-	ID         int64  `json:"id"`
-	UserID     int64  `json:"user_id"`
-	QuestionID int64  `json:"question_id"`
-	ChoiceID   int64  `json:"choice_id"`
+	ID         int    `json:"id"`
+	UserID     int    `json:"user_id"`
+	QuestionID int    `json:"question_id"`
+	ChoiceID   int    `json:"choice_id"`
 	IsRight    bool   `json:"is_right"`
 	AnswerTime string `json:"answer_time"`
 }
@@ -29,5 +29,6 @@ func (answer Answer) Validate(isRight int) rest_errors.RestErr {
 	if isRight < 0 || isRight > 1 {
 		return rest_errors.NewBadRequestError("invalid status value")
 	}
+
 	return nil
 }
